requirements.txt
Django==5.0.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==23.5
